cursus/7Ring_Pass/ft_transcendence/backend/requirements.txt
# Back-end requirements :

wheel  # A built-package format for Python. A wheel is a ZIP-format archive with a specially formatted filename and the .whl extension.
asgiref==3.6.0  # ASGI is a standard interface between async-capable Python web servers, applications, and frameworks.
async-timeout==4.0.3 # Timeout context manager for asyncio programs.
attrs==23.2.0  # attrs is the Python package that will bring back the joy of writing classes by relieving you from the drudgery of implementing object protocols (like __init__).
autobahn==23.6.2  # Autobahn|Python is a subproject of Autobahn and provides open-source implementations of the WebSocket and WAMP protocols for Python.
Automat==22.10.0  # A library for concise, idiomatic Python expression of finite-state automata (particularly deterministic finite-state transducers).
celery==5.2.3  # Celery is an asynchronous task queue/job queue based on distributed message passing.
certifi==2024.2.2  # A carefully curated collection of Root Certificates for validating the trustworthiness of SSL certificates while verifying the identity of TLS hosts.
cffi==1.16.0  # Foreign Function Interface for Python calling C code.
channels==4.1.0 # A Django extension for work with WebSockets 
channels_redis==4.2.0 # A comunication layer used by Redis for backend
charset-normalizer==3.3.2  # A library for character encoding detection, primarily aimed at web content.
constantly==23.10.4  # A library that provides symbolic constant support. It includes collections and constants with text, numeric, and bit flag values.
coreapi==2.3.3  # CoreAPI is designed to be a browsable representation of API data.
coreschema==0.0.4  # Core schema is a project providing schema generation for CoreAPI.
cryptography==42.0.7  # A package designed to expose cryptographic primitives and recipes to Python developers.
daphne==4.1.2  # Daphne is an HTTP, HTTP2, and WebSocket protocol server for ASGI and ASGI-HTTP, developed as part of the Django Channels project.
defusedxml==0.8.0rc2  # XML parsing library that prevents XML attacks in your applications.
Django==4.2.11  # A high-level Python Web framework that encourages rapid development and clean, pragmatic design.
django_celery_beat==2.6.0  # Django-Celery-Beat is a Django extension that stores the schedule in the Django database and uses the Django ORM for managing periodic tasks.
django-celery-results==2.4.0  # Django-Celery-Results is an extension to Celery which adds result backends for Django.
django-cors-headers==4.3.1  # A Django App that adds Cross-Origin Resource Sharing (CORS) headers to responses.
django-elasticsearch-dsl  # Integration of Elasticsearch DSL with Django.
django-environ==0.9.0  # Django-environ allows you to utilize 12factor inspired environment variables to configure your Django application.
django-templated-mail==1.1.1  # A Django library for sending templated emails.
django-prometheus==2.4.0.dev5 # Exports Django monitoring metrics for Prometheus.io, providing insights into requests, responses, database access, and cache operations
django-redis==5.0.0 # A Django framework library that allows it to be used as a cache.
djangorestframework==3.15.1  # A powerful and flexible toolkit for building Web APIs.
djangorestframework-simplejwt==5.3.1  # A JSON Web Token authentication plugin for the Django REST Framework.
djoser==2.2.2  # REST implementation of Django authentication system.
hyperlink==21.0.0  # A featureful, immutable, and correct URL for Python.
idna==3.7  # Internationalized Domain Names in Applications (IDNA) package for Python.
incremental==22.10.0  # A small library that versions your Python projects.
itypes==1.2.0  # Simple immutable types for Python, to construct an immutable type collection of any sort.
Jinja2==3.1.4  # A modern and designer-friendly templating language for Python.
MarkupSafe==2.1.5  # Implements a XML/HTML/XHTML Markup safe string for Python.
msgpack==1.0.8  # MessagePack (msgpack) is a binary serialization format that is fast and compact.
oauthlib==3.2.2  # A generic, spec-compliant, thorough implementation of the OAuth request-signing logic.
pillow==10.3.0 # The Python Imaging Library adds image processing capabilities to your Python interpreter.
postgres==4.0  # A wrapper for the psycopg2 library that simplifies PostgreSQL database interactions for Python applications.
psycopg2-binary==2.9.6  # The most popular PostgreSQL adapter for the Python programming language.
psycopg2-pool==1.2  # A database connection manager for psycopg2 that implements a connection pool.
pyasn1==0.6.0  # ASN.1 types and codecs (X.208).
pyasn1_modules==0.4.0  # A collection of ASN.1-based protocols modules.
pycparser==2.22  # A complete parser of the C language, written in pure Python.
PyJWT==2.8.0  # A Python library which allows you to encode and decode JSON Web Tokens (JWT).
python-logstash  # Python logging handler for Logstash.
pyOpenSSL==24.1.0  # A robust, commercial-grade, full-featured, and Open Source cryptography library for Python.
python3-openid==3.2.0  # A set of Python packages to support use of the OpenID decentralized identity system in your application.
pytz==2021.3  # World timezone definitions, modern and historical. Brings the Olson tz database into Python.
redis==5.0.4  # Redis is an in-memory data structure store, used as a database, cache, and message broker.
requests==2.31.0  # A simple, yet elegant HTTP library.
requests-oauthlib==2.0.0  # OAuthlib authentication support for Requests.
service-identity==24.1.0  # Service identity verification for pyOpenSSL.
six==1.16.0  # Python 2 and 3 compatibility utilities.
social-auth-app-django==5.4.1  # Django social authentication made simple.
social-auth-core==4.5.4  # Core functionalities for the social-auth ecosystem.
sqlparse==0.4.3  # A non-validating SQL parser module for Python.
Twisted==24.3.0  # An event-driven networking engine written in Python.
txaio==23.1.1  # Compatibility API between asyncio/Twisted/Trollius.
typing_extensions==4.11.0  # Backported and Experimental Type Hints for Python 3.5+.
uritemplate==4.1.1  # URI templates as described by RFC 6570.
urllib3==2.2.1  # A powerful, sanity-friendly HTTP client for Python.
zope.interface==6.4.post1  # A framework for defining interfaces in Python.
